fix(IndicadorConverter): resolvido problema que fazia a devoluçao em PDV ir zerada

+++ src/main/java/br/com/zalf/prolog/webservice/entrega/indicador/IndicadorConverter.java
@@ -69,8 +69,8 @@ public class IndicadorConverter {
             DevPdv item = new DevPdv();
             item.setData(rSet.getDate("DATA"))
                     .setMapa(rSet.getInt("MAPA"))
-                    .setOk(rSet.getInt("ENTREGASCOMPLETAS") + rSet.getInt("ENTREGASNAOREALIZADAS"))
-                    .setNok(rSet.getInt("ENTREGASNAOREALIZADAS"))
+                    .setOk(rSet.getDouble("ENTREGASCOMPLETAS") + rSet.getInt("ENTREGASNAOREALIZADAS"))
+                    .setNok(rSet.getDouble("ENTREGASNAOREALIZADAS"))
                     .setMeta(rSet.getDouble("META_DEV_PDV"))
                     .calculaResultado();
         return item;
